+++ core/src/backgammon.py
@@ -6,12 +6,13 @@ from enum import Enum, auto
 from typing import List # Para el type hinting
 import random # Para la tirada inicial
 from core.src.exceptions import (
-    PosicionOcupadaException, 
-    PrimerCuadranteIncompletoException, 
+    PosicionOcupadaException,
+    PrimerCuadranteIncompletoException,
     PosicionVaciaException,
     MovimientoNoPosibleException,
     TiradaInicialEmpateException,
-    NoEsMomentoException
+    NoEsMomentoException,
+    FichaContrariaException
 )
 
 class EstadoJuego(Enum):
@@ -21,7 +22,8 @@ class EstadoJuego(Enum):
     FIN_JUEGO = auto()
 
 class BackgammonGame:
-    """Clase principal que maneja la lógica del juego de Backgammon
+    """
+    Clase principal que maneja la lógica del juego de Backgammon
 
     Atributos
     ---------
@@ -79,9 +81,19 @@ class BackgammonGame:
         """Devuelve el ganador si es que lo hay"""
         return self.__ganador__
 
+    @property
+    def jugador1(self) -> Jugador:
+        """Devuelve el jugador 1"""
+        return self.__jugador1__
+
+    @property
+    def jugador2(self) -> Jugador:
+        """Devuelve el jugador 2"""
+        return self.__jugador2__
+
     def lanzar_dados(self) -> None:
         """Tira los dados, si es el primer turno, decide quien empieza, si no, es una tirada normal.
-        
+
         Raises
         ------
         NoEsMomentoException
@@ -90,7 +102,7 @@ class BackgammonGame:
             Se dio un empate en la tirada inicial
         """
 
-        if self.__estado_juego__ != EstadoJuego.TIRANDO_DADOS:
+        if self.estado_actual != EstadoJuego.TIRANDO_DADOS:
             raise NoEsMomentoException('No es el momento de tirar los dados')
 
         if self.__primer_turno__ == True:
@@ -100,9 +112,9 @@ class BackgammonGame:
             if dado_j1 == dado_j2:
                 raise TiradaInicialEmpateException('Empate, hay que tirar de nuevo')
             elif dado_j1 > dado_j2:
-                self.__jugador_actual__ = self.__jugador1__
+                self.__jugador_actual__ = self.jugador1
             else:
-                self.__jugador_actual__ = self.__jugador2__
+                self.__jugador_actual__ = self.jugador2
 
             self.__dados__.set_valores([dado_j1, dado_j2])
             self.__primer_turno__ = False
@@ -111,52 +123,103 @@ class BackgammonGame:
             return {
                 'dado_j1': dado_j1,
                 'dado_j2': dado_j2,
-                'ganador': self.__jugador_actual__.__nombre__
+                'ganador': self.turno_actual.nombre
             }
         else:
             self.__dados__.tirar()
             self.__estado_juego__ = EstadoJuego.MOVIENDO
             return {
-                'dados': self.__dados__.valores
+                'dados': self.dados
             }
+
+    def intentar_mover_ficha(self, posicion: int, dado_usado: int) -> None:
+        """Intenta mover una ficha desde una posición del tablero a otra, o a fuera del tablero
+
+        Parametros
+        ----------
+        posicion : int
+            La posición de origen de la ficha que se intenta mover (1-24)
+        dado_usado : int
+            El valor del dado que se desea usar
+
+        Raises
+        ------
+        NoEsMomentoException
+            Si el estado del juego no es moviendo
+        ValueError
+            Si el valor de la posicion no es válido (entre 1 y 24)
+            Si el dado no está disponible
+        Exception
+            Si tienes fichas en la barra
+        FichaContrariaException
+            Si intentas mover una ficha del rival
+        """
+
+        if not (1 <= posicion <= 24):
+            raise ValueError('La posición de origen debe estar entre 1 y 24')
+
+        if self.estado_actual != EstadoJuego.MOVIENDO:
+            raise NoEsMomentoException('No es el momento de mover fichas')
+
+        if dado_usado not in self.dados:
+            raise ValueError(f'El dado {dado_usado} no está disponible')
+
+        turno_color = self.turno_actual.color
+        if not self.tablero.barra_vacia(turno_color):
+            raise Exception('Tienes fichas en la barra')
+
+        try:
+            if not self.tablero.get_ficha(posicion, turno_color):
+                raise FichaContrariaException('Esa ficha no es tuya')
+
+            self.tablero.mover(posicion, dado_usado, turno_color)
+            self.__dados__.usar(dado_usado)
+            self.comprobar_estado_post_movimiento()
+
+        except (
+            PosicionVaciaException,
+            PosicionOcupadaException,
+            PrimerCuadranteIncompletoException
+        ):
+            raise
 
     def cambiar_turno(self) -> None:
         """Pasa el turno al siguiente jugador y resetea los dados"""
-        if self.__jugador_actual__ == self.__jugador1__:
-            self.__jugador_actual__ = self.__jugador2__
+        if self.turno_actual == self.jugador1:
+            self.__jugador_actual__ = self.jugador2
         else:
-            self.__jugador_actual__ = self.__jugador1__
+            self.__jugador_actual__ = self.jugador1
 
         self.__estado_juego__ = EstadoJuego.TIRANDO_DADOS
         self.__dados__.resetear()
-        print(f'Turno de {self.__jugador_actual__.__nombre__}')
+        print(f'Turno de {self.turno_actual.nombre}')
 
     def saltar_turno(self) -> None:
         """Permite al jugador saltar el turno si no le quedan movimientos válidos (Usado por el controlador)
-        
+
         Raises
         ------
         NoEsMomentoException
             El estado del juego no es moviendo
         """
-        if self.__estado_juego__ != EstadoJuego.MOVIENDO:
+        if self.estado_actual != EstadoJuego.MOVIENDO:
             raise NoEsMomentoException('No puedes saltar turno si no estás moviendo fichas')
-        print(f'{self.__jugador_actual__.__nombre__} salta el turno')
+        print(f'{self.turno_actual.nombre} salta el turno')
         self.cambiar_turno()
 
     def comprobar_estado_post_movimiento(self) -> None:
         """Se llama después de cada movimiento para ver si el juego ha terminado o si se han agotado los dados"""
-        turno_color = self.__jugador_actual__.__color__
+        turno_color = self.turno_actual.color
 
-        if self.__tablero__.condicion_victoria(turno_color):
+        if self.tablero.condicion_victoria(turno_color):
             self.__estado_juego__ = EstadoJuego.FIN_JUEGO
-            self.__ganador__ = self.__jugador_actual__
+            self.__ganador__ = self.turno_actual
             return
 
         if not self.__dados__.dados_disponibles():
             self.cambiar_turno()
 
-            
 
-    
+
+